Skip O0 LTO in local test script

+++ scripts/local_test.sh
@@ -5,11 +5,14 @@ for platform in "arm" "thumb" "aarch64"; do
   for opt_level in "0" "3" "s"; do
     for sanitizers in "ON" "OFF"; do
       for lto in "ON" "OFF"; do
-        rm -f CMakeCache.txt
-        cmake . -DBUILD_PLATFORM=${platform} -DOPT_LEVEL=${opt_level} -DSANITIZERS=${sanitizers} -DLTO=${lto}
-        make make_demos
-        lit demos/ -a
-        make clean
+        # LTO isn't enabled at O0
+        if [[ "${opt_level}_${lto}" != "0_ON" ]]; then
+          rm -f CMakeCache.txt
+          cmake . -DBUILD_PLATFORM=${platform} -DOPT_LEVEL=${opt_level} -DSANITIZERS=${sanitizers} -DLTO=${lto}
+          make -j2 make_demos
+          lit demos/ -a
+          make clean
+        fi
       done
     done
   done
